+++ sfja/src/components/FormManager/BlankFormBuilder/FormBuilder/star-rating.jsx
