+++ sfja/src/components/FormManager/BlankFormBuilder/scss/application.scss
