+++ sfja/src/components/FormManager/BlankFormBuilder/scss/react-bootstrap-slider.scss
@@ -1,0 +1,3 @@
+$slider-horizontal-width: 100%;
+
+@import "../../../../../node_modules/bootstrap-slider/src/sass/bootstrap-slider";
