+++ sfja/src/components/FormManager/BlankFormBuilder/scss/react-star-rating.scss
