+++ sfja/src/components/FormManager/BlankFormBuilder/scss/variables.scss
