--- sfja/src/components/FormManager/scss/react-bootstrap-slider.scss
@@ -1,3 +1,0 @@
-$slider-horizontal-width: 100%;
-
-@import "../sfja/node_modules/bootstrap-slider/src/sass/bootstrap-slider";
