+++ sfja/src/components/LoginPage.css
@@ -75,7 +75,7 @@ p {
     margin-top: 20px;
     margin-bottom: -10px;
     font-size: 32px;
-    color: #1F3F7F;
+    color: SteelBlue;
     font-weight: 500;
     font-family: Futura;
 }
