+++ sfja/src/components/ParentDash/FormDisplay.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import ReactFormGenerator from '../FormManager/FormBuilder/form';
+// eslint-disable-next-line max-len
+import ReactFormGenerator from '../FormManager/BlankFormBuilder/FormBuilder/form';
 
 // eslint-disable-next-line require-jsdoc
 class FormDisplay extends React.Component {
